refactor: consolidate backend utilities
